Infrastruktura/webapp/requirements.txt
flask
requests
azure-identity
gunicorn
